Added a timeout condition(customizable) to avoid infinite loops in case the input numbers will (likely) never meet.
Changed "int" values to "ulong" and "uint" to allow usage of bigger numbers.

+++ RatioGetterLibrary/Number.cs
@@ -5,20 +5,20 @@ using System.Text;
 namespace RatioGetterLibrary {
     public class Number {
 
-        public Number(string name, int baseValue, int? timeout = null) {
+        public Number(string name, uint baseValue, uint? timeout = null) {
             Name = name;
             BaseValue = baseValue;
             this.timeout = timeout;
         }
 
         public string Name { get; private set; }
-        public int BaseValue { get; }
-        public int Multiplier { get; private set; } = 1;
-        public int Result { get => BaseValue * Multiplier; }
+        public uint BaseValue { get; }
+        public ulong Multiplier { get; private set; } = 1;
+        public ulong Result { get => BaseValue * Multiplier; }
         public bool isTimeout { get => (timeout != null) && timeout <= Multiplier; }
-        private int? timeout { get; }
+        private uint? timeout { get; }
 
-        public void Next(int addValue = 1) {
+        public void Next(ulong addValue = 1) {
             Multiplier += addValue;
         }
         public override string ToString() {
